requirements.txt
bleach==2.0.0
branca==0.2.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
Django==2.0.7
django-tastypie==0.14.1
entrypoints==0.2.3
folium==0.3.0
glob2==0.5
html5lib==0.999999999
idna==2.7
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
nine==1.0.0
notebook==5.0.0
olefile==0.44
pandocfilters==1.4.2
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
Pygments==2.2.0
pyparsing==2.2.0
pyserial==3.3
python-dateutil==2.6.1
python-decouple==3.1
python-mimeparse==1.6.0
python-utils==2.2.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.19.1
simplegeneric==0.8.1
six==1.10.0
stl==0.0.3
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.23
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
widgetsnbextension==3.0.0
xlrd==1.0.0
xlwt==1.2.0
